+++ CHANGES.rst
@@ -98,6 +98,15 @@ Fixes
 - Adapted a query for compatibility with the upcoming CrateDB 4.0 release.
 
 
+2019/06/30 1.15.1
+=================
+
+Feature
+-------
+
+- sharding-calculator: a simple tool to walk a user through the process, how
+  to split their crate table into shards. Can also read stats from existing table.
+
 2019/04/03 1.15.0
 =================